Created audio folder, created Yggdrasil, new onload function
I've changed a bit how desktop icons are created and handled. Now it's done by Yggdrasil, a global array that contains all directories, files and archives.

New classes for all of them, meaning: `CDirectory`, `CArchive` that extends `CDirectory`  and CFile. Inside main `script.js` now there is a function `loadYggdrasil` that is the first in the call order. It populates the global `FOLDERS` and `ARCHIVES` arrays. Why? Because I've said so.

AudioPlayer is on the way